Move getQueryTimelineName into runtime utils
