+++ playground/composables/useUrqlClient.ts
@@ -1,7 +1,6 @@
 import type { DocumentNode } from 'graphql'
 import { provideClient, useQuery } from '@urql/vue'
 import type { AnyVariables } from '@urql/vue'
-import { getQueryTimelineName } from '../utils'
 
 export const URQL_CLIENT_NUXT_APP_KEY = 'urql'
 
@@ -16,6 +15,8 @@ export function useUrqlQuery<T = any, V extends AnyVariables = AnyVariables>(
   const { query, variables } = options
 
   const timeline = useRequestTimeline()
+  const { getQueryTimelineName } = useRequestTimelineUtils()
+
   const client = useUrqlClient()
   provideClient(client)
