+++ src/runtime/composables.ts
@@ -1,6 +1,10 @@
 import { useNuxtApp } from '#imports'
-import { getRequestTimeline } from './utils.js'
+import { getRequestTimeline, getQueryTimelineName } from './utils.js'
 
 export function useRequestTimeline() {
   return getRequestTimeline(useNuxtApp())
+}
+
+export function useRequestTimelineUtils() {
+  return { getQueryTimelineName }
 }
