+++ src/runtime/utils.ts
@@ -1,3 +1,4 @@
+import type { DocumentNode } from 'graphql'
 import type { NuxtApp } from '#app'
 import {
   REQUEST_TIMELINE_KEY,
@@ -5,6 +6,10 @@ import {
   REQUEST_TIMELINE_ROUTE_QUERY,
 } from '../constants'
 import type { RequestTimeline } from '../types'
+
+type AnyObject = Record<string, unknown>
+
+const MAX_LENGTH_QUERY_VARS_TIMELINE = 60
 
 export function getRequestTimeline(nuxtApp: NuxtApp) {
   return nuxtApp[`$${REQUEST_TIMELINE_KEY}`]
@@ -23,4 +28,38 @@ export function generateUrl(requestChunks?: RequestTimeline['requestChunks']): s
   url.searchParams.append(REQUEST_TIMELINE_ROUTE_QUERY, JSON.stringify(requestChunks))
 
   return url.toString()
+}
+
+/**
+ * Generate name to display on the table of the RequestTimeline page.
+ */
+export function getQueryTimelineName<QueryVariables>(options: {
+  query: DocumentNode
+  variables: QueryVariables
+  excludedVars?: string[]
+}): string {
+  const { query, variables, excludedVars = [] as string[] } = options
+
+  const queryVars = {} as AnyObject
+  for (const varKey in variables) {
+    if (!excludedVars.includes(varKey)) queryVars[varKey] = variables[varKey]
+  }
+
+  return query.definitions
+    ?.map(def => {
+      if ('operation' in def && 'name' in def) {
+        let name = def.operation === 'query' ? (def.name || {}).value || '' : ''
+        if (Object.keys(queryVars).length) name += ` ${stringifyQueryVariables(queryVars)}`
+        return name
+      }
+    })
+    .filter(v => v)
+    .join(',')
+}
+
+function stringifyQueryVariables(variables: AnyObject) {
+  const strVars = JSON.stringify(variables)
+  return strVars.length > MAX_LENGTH_QUERY_VARS_TIMELINE
+    ? `${strVars.substring(0, MAX_LENGTH_QUERY_VARS_TIMELINE)}...}`
+    : strVars
 }
